Extract lamdera/container extensions into TH loaded file and ensure its always loaded in Lamdera contexts
